add atomsk.pl

+++ packages/mpich4GCC_devices.pl
@@ -25,16 +25,17 @@ sub ld_setting {
 	my $ld_library_path = $ENV{'LD_LIBRARY_PATH'};	
 	$ENV{'LD_LIBRARY_PATH'} = "$attached_ld:$ld_library_path";		
 }
-#my $mattached_path = "/opt/mpich-3.3.2/bin";#attached path in main script
-#path_setting($mattached_path);
+my $mattached_path = "/opt/nvidia/hpc_sdk/Linux_x86_64/21.2/compilers/bin";#attached path in main script
+path_setting($mattached_path);
 my $mattached_ld = "/usr/local/lib";#attached ld path in main script
 ld_setting($mattached_ld);
 
 use warnings;
 use strict;
-use Env::Modify qw(:sh source);
 use Cwd; #Find Current Path
 use File::Copy; # Copy File
+
+my $wgetORgit = "yes";
 
 my $packageDir = "/home/packages";
 if(!-e $packageDir){# if no /home/packages, make this folder	
@@ -48,23 +49,26 @@ my $thread4make = `lscpu|grep "^CPU(s):" | sed 's/^CPU(s): *//g'`;
 chomp $thread4make;
 print "Total threads can be used for make: $thread4make\n";
 
-my $currentVer = "mpich-3.4.1";#***** the latest version of this package
-my $prefixPath = "/opt/$currentVer";#you may use your own
+my $currentVer = "mpich-3.3.2";#***** the latest version of this package
+my $prefixPath = "/opt/pgi-$currentVer";#you may use your own
 system ("rm -rf $prefixPath");# remove the older directory first
-my $URL = "http://www.mpich.org/static/downloads/3.4.1/mpich-3.4.1.tar.gz";
-my $Dir4download = "$packageDir/$currentVer"."_download"; #the directory we download MPICH
+my $URL = "http://www.mpich.org/static/downloads/3.3.2/mpich-3.3.2.tar.gz";
+my $Dir4download = "$packageDir/pgi-$currentVer"."_download"; #the directory we download MPICH
 
 ###
-system ("rm -rf $Dir4download");# remove the older directory first
-system("mkdir $Dir4download");# make a directory in current path
-
-chdir("$Dir4download");# cd to this dir for downloading the packages
-#get the latest package in the directory and save it as the filename you want
-system("wget $URL"); 
-if ($?){die "wget -O $currentVer $URL failed\n";}
-if(! (-e "$Dir4download/$currentVer.tar.gz")){die "No $currentVer downloaded";}# if no mpich file
-
+if($wgetORgit eq "yes"){
+	system ("rm -rf $Dir4download");# remove the older directory first
+	system("mkdir $Dir4download");# make a directory in current path
+	
+	chdir("$Dir4download");# cd to this dir for downloading the packages
+	#get the latest package in the directory and save it as the filename you want
+	system("wget $URL"); 
+	if ($?){die "wget -O $currentVer $URL failed\n";}
+	if(! (-e "$Dir4download/$currentVer.tar.gz")){die "No $currentVer downloaded";}# if no mpich file
+}
 ## tar -xvzf XXX(package name), and then cd this new folder	
+chdir("$Dir4download");# cd to this dir for downloading the packages
+
 system("rm -rf  $currentVer");
 system("tar -xvzf $currentVer.tar.gz"); #$Ch =  Check
 if ($?){die "tar -xvzf mpich failed\n";} 
@@ -73,7 +77,9 @@ chdir("$Dir4download/$currentVer");#$currentVer is the directory name after tar
 unlink "Makefile";
 sleep(1);
 #--enable-fast=all,O3 --with-slurm-include=/usr/local/include/slurmCPPFLAGS=-I/home/packages/mpich_download/mpich-3.3.2/src/pmi/pmi2/include
-system("./configure --prefix=$prefixPath --with-device=ch4:ucx --enable-fast=all,O3");# --with-slurm=[/usr/local] VERBOSE=1 |tee 00mpich_configure.txt"); #./configure
+#system("./configure --prefix=$prefixPath --with-device=ch4:ucx --enable-fast=all,O3");# --with-slurm=[/usr/local] VERBOSE=1 |tee 00mpich_configure.txt"); #./configure
+system("./configure   CC=pgcc F77=pgf77 FC=pgf90 CXX=pgc++ --prefix=$prefixPath --disable-builtin-atomics --with-device=ch4:ucx --enable-fast=all,O3 ");# --with-slurm=[/usr/local] VERBOSE=1 |tee 00mpich_configure.txt"); #./configure
+#CC=pgcc F77=pgf77 FC=pgf90 CXX=pgc++
 if($?){die "config $currentVer failed!\nReason $?:$!\n";}
 
 ##after the configure process is done, type "make" and then "make install"
